+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+## [1.10.0] - 2017-08-03
 ### Added
+- `be.objValueInArray`, check if is a key of an object with determined value is in array
 - `be.generatorFunction`, check if is GeneratorFunction
 
 ## [1.9.0] - 2017-07-30